+++ test/ex_acme_test.exs
@@ -28,4 +28,22 @@ defmodule ExAcmeTest do
 
     assert external_account_required == false
   end
+
+  test "named directory lets_encrypt" do
+    {:ok, pid} = ExAcme.start_link(finch: MyFinch, directory_url: :lets_encrypt)
+    %{directory_url: directory_url} = Agent.get(pid, & &1)
+    assert directory_url == "https://acme-v02.api.letsencrypt.org/directory"
+  end
+
+  test "named directory lets_encrypt_staging" do
+    {:ok, pid} = ExAcme.start_link(finch: MyFinch, directory_url: :lets_encrypt_staging)
+    %{directory_url: directory_url} = Agent.get(pid, & &1)
+    assert directory_url == "https://acme-staging-v02.api.letsencrypt.org/directory"
+  end
+
+  test "named directory zerossl" do
+    {:ok, pid} = ExAcme.start_link(finch: MyFinch, directory_url: :zerossl)
+    %{directory_url: directory_url} = Agent.get(pid, & &1)
+    assert directory_url == "https://acme.zerossl.com/v2/DV90"
+  end
 end